+++ lib/src/data_transfer_objects/transaction/transfer_payload.dart
@@ -15,7 +15,7 @@ class TransferPayload with _$TransferPayload {
     @JsonKey(name: 'commitment') required List<int> commitment,
     @JsonKey(name: 'ct_validity_proof')
     required Map<String, dynamic> validityProof,
-    @JsonKey(name: 'destination') required List<int> destination,
+    @JsonKey(name: 'destination') required dynamic destination,
     @JsonKey(name: 'receiver_handle') required List<int> receiverHandle,
     @JsonKey(name: 'sender_handle') required List<int> senderHandle,
     @JsonKey(name: 'extra_data') dynamic extraData,