+++ lib/src/data_transfer_objects/transaction/transfer_payload.g.dart
@@ -16,9 +16,7 @@ _$TransferPayloadImpl _$$TransferPayloadImplFromJson(
           .map((e) => (e as num).toInt())
           .toList(),
       validityProof: json['ct_validity_proof'] as Map<String, dynamic>,
-      destination: (json['destination'] as List<dynamic>)
-          .map((e) => (e as num).toInt())
-          .toList(),
+      destination: json['destination'],
       receiverHandle: (json['receiver_handle'] as List<dynamic>)
           .map((e) => (e as num).toInt())
           .toList(),