+++ lib/src/data_transfer_objects/wallet/build_unsigned_transaction/unsigned_transaction_response.dart
@@ -30,7 +30,7 @@ class UnsignedTransactionResponse with _$UnsignedTransactionResponse {
   factory UnsignedTransactionResponse.fromJson(Map<String, dynamic> json) {
     if (json
         case {
-          'data': final Map<String, dynamic> transfers,
+          'data': {'burn': final Map<String, dynamic> burn},
           'fee': final int fee,
           'hash': final String hash,
           'version': final int version,
@@ -42,7 +42,7 @@ class UnsignedTransactionResponse with _$UnsignedTransactionResponse {
           'threshold': final int threshold,
         }) {
       return UnsignedTransactionResponse(
-        data: Transfers.fromJson(transfers),
+        data: BurnPayload.fromJson(burn),
         fee: fee,
         hash: hash,
         version: version,
@@ -61,7 +61,7 @@ class UnsignedTransactionResponse with _$UnsignedTransactionResponse {
       );
     } else if (json
         case {
-          'data': {'burn': final Map<String, dynamic> burn},
+          'data': {'multi_sig': final Map<String, dynamic> multisig},
           'fee': final int fee,
           'hash': final String hash,
           'version': final int version,
@@ -73,7 +73,38 @@ class UnsignedTransactionResponse with _$UnsignedTransactionResponse {
           'threshold': final int threshold,
         }) {
       return UnsignedTransactionResponse(
-        data: Burn.fromJson(burn),
+        data: MultisigPayload.fromJson(multisig),
+        fee: fee,
+        hash: hash,
+        version: version,
+        nonce: nonce,
+        source: source.map((e) => e as int).toList(),
+        rangeProof: rangeProof.map((e) => e as int).toList(),
+        sourceCommitments:
+            sourceCommitments.map((e) => e as Map<String, dynamic>).toList(),
+        reference: Reference.fromJson(reference),
+        threshold: threshold,
+        txAsHex:
+            (json['tx_as_hex'] != null) ? json['tx_as_hex'] as String : null,
+        multiSig: (json['multi_sig'] != null)
+            ? Multisig.fromJson(json['multi_sig'] as Map<String, dynamic>)
+            : null,
+      );
+    } else if (json
+        case {
+          'data': final Map<String, dynamic> transfers,
+          'fee': final int fee,
+          'hash': final String hash,
+          'version': final int version,
+          'nonce': final int nonce,
+          'source': final List<dynamic> source,
+          'range_proof': final List<dynamic> rangeProof,
+          'source_commitments': final List<dynamic> sourceCommitments,
+          'reference': final Map<String, dynamic> reference,
+          'threshold': final int threshold,
+        }) {
+      return UnsignedTransactionResponse(
+        data: TransfersPayload.fromJson(transfers),
         fee: fee,
         hash: hash,
         version: version,